+++ runtime.es.js
@@ -1,0 +1,1 @@
+export { pug as default } from './dist/runtime.es.js'