+++ src/utils/assign.js
@@ -1,5 +1,5 @@
 /**
- * Object.assign like function, but converts falsy `dest` to object.
+ * Object.assign like function, but always converts falsy `dest` to object.
  *
  * @param   {any} dest - An object or falsy value
  * @returns {Object}   object with merged properties
@@ -7,13 +7,13 @@
 export default function assign (dest) {
   const args = arguments
 
-  dest = dest && Object(dest) || {}
+  dest = dest ? Object(dest) : {}
 
   for (let i = 1; i < args.length; i++) {
     const src = args[i]
 
     if (src) {
-      const keys = Object.keys(Object(src))
+      const keys = Object.keys(src)
 
       for (let j = 0; j < keys.length; j++) {
         const p = keys[j]