+++ test/fixtures/custom_globals/main.js
@@ -1,0 +1,3 @@
+import template from './template.pug'
+
+export default template()