+++ test/fixtures/runtime_import/main.js
@@ -1,0 +1,5 @@
+import template from './template.pug'
+
+var data = { name: 'World' }
+
+assert.equal(template(data), '<p>Hello World</p>')