+++ test/run.js
@@ -27,7 +27,10 @@ function executeBundle (bundle, name) {
   var result = bundle.generate({
     format: 'cjs'
   })
-  writeBundle(bundle, name)
+
+  if (name) {
+    writeBundle(bundle, name)
+  }
 
   // eslint-disable-next-line no-new-func
   var fn = new Function('require', 'module', 'assert', result.code)
@@ -35,6 +38,7 @@ function executeBundle (bundle, name) {
 
   fn(require, module, assert)
 }
+
 
 describe('rollup-plugin-pug', function () {
 
@@ -77,6 +81,58 @@ describe('rollup-plugin-pug', function () {
     .catch(done)
   })
 
+  it('allows custom globals in addition to the default ones', function (done) {
+    rollup({
+      entry: 'fixtures/custom_globals/main.js',
+      plugins: [_pug({
+        sourceMap: false,
+        globals: ['require', 'String'] // String is default, coverage only
+      })]
+    })
+    .then(function (bundle) {
+      var result = bundle.generate({
+        format: 'es'
+      })
+      writeBundle(bundle, 'custom_globals')
+
+      assert(result.code.indexOf('typeof require2!') > -1)
+      assert(result.code.indexOf('typeof require!') < 0)
+      assert(result.code.indexOf('typeof Math!') < 0)
+
+      done()
+    })
+    .catch(done)
+  })
+
+  it('can use .pug extensions', function (done) {
+    rollup({
+      entry: 'fixtures/basic/main.js',
+      plugins: [_pug({
+        extensions: 'pug'
+      })]
+    })
+    .then(function (bundle) {
+      executeBundle(bundle)
+      done()
+    })
+    .catch(done)
+  })
+
+  it('can use the `include: []` option with `extensions: "*"`', function (done) {
+    rollup({
+      entry: 'fixtures/basic/main.js',
+      plugins: [_pug({
+        include: '**/*.pug',
+        extensions: '*'
+      })]
+    })
+    .then(function (bundle) {
+      executeBundle(bundle)
+      done()
+    })
+    .catch(done)
+  })
+
   it('move imports after a dash out of the pug function', function (done) {
     rollup({
       entry: 'fixtures/imports/main.js',
@@ -104,6 +160,22 @@ describe('rollup-plugin-pug', function () {
     .catch(done)
   })
 
+  it('static compilation ignores compileDebug', function (done) {
+    rollup({
+      entry: 'fixtures/precompile/main.js',
+      plugins: [_pug({
+        locals: { name: 'pug', other: 'other' },
+        option_local: 'option_local',
+        compileDebug: true
+      })]
+    })
+    .then(function (bundle) {
+      executeBundle(bundle, 'precompile_debug')
+      done()
+    })
+    .catch(done)
+  })
+
   it('defaults to the correct path based on rollup entry', function (done) {
     rollup({
       entry: '../test/fixtures/app/index.js',
@@ -121,4 +193,71 @@ describe('rollup-plugin-pug', function () {
     })
     .catch(done)
   })
+
+  it('does not import the Pug runtime if `pugRuntime` is falsy', function (done) {
+    rollup({
+      entry: 'fixtures/no_runtime/main.js',
+      plugins: [_pug({
+        pugRuntime: false
+      })]
+    })
+    .then(function (bundle) {
+      executeBundle(bundle, 'no_runtime')
+      done()
+    })
+    .catch(done)
+  })
+
+  it('allows import a custom Pug runtime with `pugRuntime: "[custom-runtime]"`', function (done) {
+    rollup({
+      entry: 'fixtures/custom_runtime/main.js',
+      plugins: [_pug({
+        pugRuntime: 'pug-runtime',
+      })],
+      external: ['pug-runtime']
+    })
+    .then(function (bundle) {
+      executeBundle(bundle, 'custom_runtime')
+      done()
+    })
+    .catch(done)
+  })
+
+  it('can import custom runtime w/ `import pug from...` and `pugRuntime: false`', function (done) {
+    var alias = require('rollup-plugin-alias')
+
+    rollup({
+      entry: 'fixtures/runtime_import/main.js',
+      plugins: [
+        _pug({
+          pugRuntime: false
+        }),
+        alias({
+          pug_runtime: '../dist/runtime.es.js'
+        })
+      ]
+    })
+    .then(function (bundle) {
+      executeBundle(bundle, 'runtime_import')
+      done()
+    })
+    .catch(done)
+  })
+
+  it('allows to skip runtime import with the `inlineRuntimeFunctions` option', function (done) {
+    rollup({
+      entry: 'fixtures/custom_runtime/main.js',
+      plugins: [_pug({
+        inlineRuntimeFunctions: true,
+        sourceMap: false
+      })]
+    })
+    .then(function (bundle) {
+      debugger
+      executeBundle(bundle, 'inline_runtime')
+      done()
+    })
+    .catch(done)
+  })
+
 })